+++ app/src/main/java/com/example/munchbox/controller/Restaurant.kt
@@ -1,6 +1,9 @@
 package com.example.munchbox.controller
 
 // TODO: delete addMeal(s) and replace usages with API endpoints
-class Restaurant(val restaurantID : String, val name : String, var meals : Set<Meal> = setOf(), val imageID: Int? = null) {
-
-}
+data class Restaurant(
+    val restaurantID : String,
+    val name : String,
+    var meals : Set<Meal> = setOf(),
+    val imageID: Int? = null
+)
